Remove testing file
Was going to test, that each controller action calls
 require_authentication!` and `require_admin!` methods.

But now I don't think, that it's necessary to test it on every action.
It is already tested in integration tests and it's hard to imagine what
would need to happen for these methods to stop being called in some
action and keep integration tests passing.

Besides, additional tests make code harder to change.